mcp/requirements.txt
mcp[cli]
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
psycopg[binary]==3.2.13
python-dotenv==1.0.1
starlette==0.38.5
